+++ openrouteservice/src/main/java/heigit/ors/routing/RouteSearchParameters.java
@@ -49,7 +49,6 @@ public class RouteSearchParameters {
     private int _weightingMethod = WeightingMethod.FASTEST;
     private Boolean _considerTraffic = false;
     private Boolean _considerTurnRestrictions = false;
-    private double _maxSpeed = -1;
     private Polygon[] _avoidAreas;
     private int _avoidFeaturesTypes;
     private int _vehicleType = HeavyVehicleAttributes.UNKNOWN;
@@ -72,14 +71,6 @@ public class RouteSearchParameters {
             throw new Exception("Routing profile is unknown.");
 
         this._profileType = profileType;
-    }
-
-    public double getMaximumSpeed() {
-        return _maxSpeed;
-    }
-
-    public void setMaximumSpeed(double maxSpeed) {
-        _maxSpeed = maxSpeed;
     }
 
     public int getWeightingMethod() {
@@ -185,14 +176,6 @@ public class RouteSearchParameters {
             throw new ParseException(ex.getMessage(), 0);
         }
 
-        if (json.has("maximum_speed")) {
-            try {
-                _maxSpeed = json.getDouble("maximum_speed");
-            } catch (Exception ex) {
-                throw new ParameterValueException(RoutingErrorCodes.INVALID_PARAMETER_FORMAT, "maximum_speed", json.getString("maximum_speed"));
-            }
-        }
-
         if (json.has("avoid_features")) {
             String keyValue = json.getString("avoid_features");
             if (!Helper.isEmpty(keyValue)) {
@@ -206,7 +189,7 @@ public class RouteSearchParameters {
                             if (flag == 0)
                                 throw new UnknownParameterValueException(RoutingErrorCodes.INVALID_PARAMETER_VALUE, "avoid_features", featName);
 
-                            if (!AvoidFeatureFlags.isValid(_profileType, flag, featName))
+                            if (!AvoidFeatureFlags.isValid(_profileType, flag))
                                 throw new ParameterValueException(RoutingErrorCodes.INVALID_PARAMETER_VALUE, "avoid_features", featName);
 
                             flags |= flag;
@@ -267,46 +250,9 @@ public class RouteSearchParameters {
 
             if (jProfileParams.has("restrictions"))
                 jRestrictions = jProfileParams.getJSONObject("restrictions");
-            if (RoutingProfileType.isCycling(_profileType)) {
-                CyclingParameters cyclingParams = new CyclingParameters();
 
-                if (jRestrictions != null) {
-                    if (jRestrictions.has("gradient"))
-                        cyclingParams.setMaximumGradient(jRestrictions.getInt("gradient"));
-
-                    if (jRestrictions.has("trail_difficulty"))
-                        cyclingParams.setMaximumTrailDifficulty(jRestrictions.getInt("trail_difficulty"));
-                }
-
-                _profileParams = cyclingParams;
-            } else if (RoutingProfileType.isWalking(_profileType)) {
-                WalkingParameters walkingParams = new WalkingParameters();
-
-                // To make the new API compatible with a new one, we create 'weightings' element.
-				/*if (!jProfileParams.has("weightings") && (jProfileParams.has("difficulty_level") || jProfileParams.has("maximum_gradient")))
-				{
-					JSONObject jWeightings = new JSONObject();
-
-					if (jProfileParams.has("difficulty_level"))
-						jWeightings.put("difficulty_level", jProfileParams.get("difficulty_level"));
-					else if	(jProfileParams.has("maximum_gradient"))
-						jWeightings.put("maximum_gradient", jProfileParams.get("maximum_gradient"));
-
-					jProfileParams.put("weightings", jWeightings);
-				}*/
-
-                if (jRestrictions != null) {
-                    if (jRestrictions.has("gradient"))
-                        walkingParams.setMaximumGradient(jRestrictions.getInt("gradient"));
-
-                    if (jRestrictions.has("trail_difficulty"))
-                        walkingParams.setMaximumTrailDifficulty(jRestrictions.getInt("trail_difficulty"));
-                }
-
-                _profileParams = walkingParams;
-            } else if (RoutingProfileType.isHeavyVehicle(_profileType)) {
+            if (RoutingProfileType.isHeavyVehicle(_profileType) == true) {
                 VehicleParameters vehicleParams = new VehicleParameters();
-
 
                 if (json.has("vehicle_type")) {
                     String vehicleType = json.getString("vehicle_type");