+++ openrouteservice/src/main/java/heigit/ors/routing/parameters/ProfileParameters.java
@@ -21,21 +21,7 @@ import java.util.List;
 
 public class ProfileParameters {
     protected ProfileWeightingCollection _weightings;
-    protected int _maximumGradient = -1;
-    
-    public ProfileParameters()
-    {
-    	
-    }
 
-    public int getMaximumGradient() {
-        return _maximumGradient;
-    }
-
-    public void setMaximumGradient(int maximumGradient) {
-        this._maximumGradient = maximumGradient;
-    }
-    
     public void add(ProfileWeighting weighting)
     {
     	if (_weightings == null)
@@ -56,7 +42,6 @@ public class ProfileParameters {
 
     public List<String> getValidRestrictions() {
         List<String> valid = new ArrayList<>();
-        valid.add("gradient");
         return valid;
     }
 }