+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Modified smoothing and buffer distances for small isochrones, aswell as other fixes for smaller isochrones (Issue #382)
 - Updated pom to use correct opengeo repo and reordered so this is the last in the list, and use latest ORS-Graphhopper library (Issue #398)
 - Added /directions as an endpoint for routing (Issue #384)
+- Removed the following avoid features: pavedroads, unpavedroads, tunnels, tracks and hills, as well as the option to set maximum speed; for cycling and walking profiles the option to specify difficulty settings such as fitness level and maximum steepness (issue #396)
 ### Deprecated
 
 ## [4.7.2] - 2018-12-10